argo: use dynamically allocated links for pairs of domains.
Use two hashtables for storing links to enable a deadlock avoidance
scheme where the two locks that protect a link can be acquired in
order of sorted domid.

Rings registered to self are stored in a link that is not shared
with any other domain. Wildcard rings are stored in another
dedicated link that is not shared with any other domain.