arm setup: use common integer-typed bootmod definition
This change enables inclusion of <xen/bootinfo.h> in Arm builds,
required for subsequent patches in this series.

It replaces the enum definition of bootmodule_kind with bootmod_type_t
as enums are not fixed size as needed for packed structs which are
needed for boot structures in subsequent changes.

Signed-off-by: Christopher Clark <christopher.w.clark@gmail.com>
Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>

New for v2 series.